+++ index.js
@@ -1,16 +1,11 @@
-var Promise = require('rsvp').Promise;
+module.exports = async function promiseMapSeries(array, iterator, thisArg) {
+  const length = array.length;
+  const cb = arguments.length > 2 ? iterator.bind(thisArg) : iterator;
+  const results = new Array(length);
 
-module.exports = function promiseMapSeries(array, iterator, thisArg) {
-  var length = array.length
-  var current = Promise.resolve()
-  var results = new Array(length)
-  var cb = arguments.length > 2 ? iterator.bind(thisArg) : iterator
-
-  for (var i = 0; i < length; ++i) {
-    current = results[i] = current.then(function(i) {
-      return cb(array[i], i, array)
-    }.bind(undefined, i))
+  for (let i = 0; i < length; ++i) {
+    results[i] = await cb(array[i], i, array);
   }
 
-  return Promise.all(results)
+  return results;
 }